Formatting away!

+++ src/main/java/org/getspout/api/util/map/TIntPairDoubleHashMap.java
@@ -1,40 +1,43 @@
 package org.getspout.api.util.map;
+
 import gnu.trove.TDoubleCollection;
 import gnu.trove.iterator.TLongDoubleIterator;
 import gnu.trove.map.hash.TLongDoubleHashMap;
 import gnu.trove.set.TLongSet;
 
 /**
- * A simplistic map that supports a pair of integers for keys, using a trove long double hashmap in the backend.
+ * A simplistic map that supports a pair of integers for keys, using a trove
+ * long double hashmap in the backend.
+ *
  * @author Afforess
  *
  */
-public class TIntPairDoubleHashMap{
+public class TIntPairDoubleHashMap {
 	private TLongDoubleHashMap map;
-	
+
 	public TIntPairDoubleHashMap() {
 		map = new TLongDoubleHashMap(100);
 	}
-	
-	public TIntPairDoubleHashMap(int capacity){
+
+	public TIntPairDoubleHashMap(int capacity) {
 		map = new TLongDoubleHashMap(capacity);
 	}
-	
+
 	public double put(int key1, int key2, double value) {
-		long key = (((long)key1)<<32) | (((long)key2) & 0xFFFFFFFFL);
+		long key = (long) key1 << 32 | key2 & 0xFFFFFFFFL;
 		return map.put(key, value);
 	}
-	
+
 	public double get(int key1, int key2) {
-		long key = (((long)key1)<<32) | (((long)key2) & 0xFFFFFFFFL);
+		long key = (long) key1 << 32 | key2 & 0xFFFFFFFFL;
 		return map.get(key);
 	}
-	
+
 	public boolean containsKey(int key1, int key2) {
-		long key = (((long)key1)<<32) | (((long)key2) & 0xFFFFFFFFL);
+		long key = (long) key1 << 32 | key2 & 0xFFFFFFFFL;
 		return map.containsKey(key);
 	}
-	
+
 	public void clear() {
 		map.clear();
 	}
@@ -44,7 +47,7 @@ public class TIntPairDoubleHashMap{
 	}
 
 	public boolean increment(int key1, int key2) {
-		long key = (((long)key1)<<32) | (((long)key2) & 0xFFFFFFFFL);
+		long key = (long) key1 << 32 | key2 & 0xFFFFFFFFL;
 		return map.increment(key);
 	}
 
@@ -65,7 +68,7 @@ public class TIntPairDoubleHashMap{
 	}
 
 	public double remove(int key1, int key2) {
-		long key = (((long)key1)<<32) | (((long)key2) & 0xFFFFFFFFL);
+		long key = (long) key1 << 32 | key2 & 0xFFFFFFFFL;
 		return map.remove(key);
 	}
