+++ src/main/java/org/getspout/api/util/map/TIntPairObjectHashMap.java
@@ -1,41 +1,44 @@
 package org.getspout.api.util.map;
+
+import java.util.Collection;
+
 import gnu.trove.iterator.TLongObjectIterator;
 import gnu.trove.map.hash.TLongObjectHashMap;
 import gnu.trove.set.TLongSet;
 
-import java.util.Collection;
-
 /**
- * A simplistic map that supports a pair of integers for keys, using a trove long object hashmap in the backend.
+ * A simplistic map that supports a pair of integers for keys, using a trove
+ * long object hashmap in the backend.
+ *
  * @author Afforess
  *
  */
-public class TIntPairObjectHashMap<K>{
+public class TIntPairObjectHashMap<K> {
 	private TLongObjectHashMap<K> map;
-	
+
 	public TIntPairObjectHashMap() {
 		map = new TLongObjectHashMap<K>(100);
 	}
-	
-	public TIntPairObjectHashMap(int capacity){
+
+	public TIntPairObjectHashMap(int capacity) {
 		map = new TLongObjectHashMap<K>(capacity);
 	}
-	
+
 	public K put(int key1, int key2, K value) {
-		long key = (((long)key1)<<32) | (((long)key2) & 0xFFFFFFFFL);
+		long key = (long) key1 << 32 | key2 & 0xFFFFFFFFL;
 		return map.put(key, value);
 	}
-	
+
 	public K get(int key1, int key2) {
-		long key = (((long)key1)<<32) | (((long)key2) & 0xFFFFFFFFL);
+		long key = (long) key1 << 32 | key2 & 0xFFFFFFFFL;
 		return map.get(key);
 	}
-	
+
 	public boolean containsKey(int key1, int key2) {
-		long key = (((long)key1)<<32) | (((long)key2) & 0xFFFFFFFFL);
+		long key = (long) key1 << 32 | key2 & 0xFFFFFFFFL;
 		return map.containsKey(key);
 	}
-	
+
 	public void clear() {
 		map.clear();
 	}
@@ -61,7 +64,7 @@ public class TIntPairObjectHashMap<K>{
 	}
 
 	public K remove(int key1, int key2) {
-		long key = (((long)key1)<<32) | (((long)key2) & 0xFFFFFFFFL);
+		long key = (long) key1 << 32 | key2 & 0xFFFFFFFFL;
 		return map.remove(key);
 	}
